:recycle: refactor(project): Moving data to test
