+++ sonar-project.properties
@@ -25,7 +25,4 @@ sonar.coverage.exclusions=\
     src/components/organisms/GeoMap/GeoMapDemo.tsx, \
     **/src/assets/**/*.json, \
     src/utils/Stickify.ts, \
-    src/test/mock-data/credit_history_data.ts, \
-    src/test/mock-data/map_dashboard_data.ts, \
-    src/test/mock-data/credit_history_data.ts, \
-    src/test/mock-data/map_dashboard_data.ts
+    src/test/mock-data/*.ts, \
